chore(desktop): rename parsecap helper to kebab-case
