+++ apps/desktop/src/lib/parse-cap.ts
